+++ shared/main/java/net/ccbluex/liquidbounce/utils/render/shader/shaders/BackgroundShader.java
@@ -35,7 +35,7 @@ public final class BackgroundShader extends Shader
 	{
 		final int resolutionID = getUniform("iResolution");
 		if (resolutionID > -1)
-			GL20.glUniform2f(resolutionID, (float) Display.getWidth(), (float) Display.getHeight());
+			GL20.glUniform2f(resolutionID, Display.getWidth(), Display.getHeight());
 		final int timeID = getUniform("iTime");
 		if (timeID > -1)
 			GL20.glUniform1f(timeID, time);
